Update test dtype

+++ tests/ops/test_simple_gla.py
@@ -110,7 +110,7 @@ def parallel_simple_gla_ref(
 @pytest.mark.parametrize('H', test_h_list)
 @pytest.mark.parametrize('D', test_d_list)
 @pytest.mark.parametrize('gate_logit_normalizer', test_gate_list)
-@pytest.mark.parametrize('dtype', [torch.float])
+@pytest.mark.parametrize('dtype', [torch.float16])
 @pytest.mark.parametrize('scale', [1, 0.1])
 @pytest.mark.skipif(
     os.getenv('SKIP_TEST_CHUNK_VARLEN') == '0',
@@ -257,7 +257,7 @@ def test_chunk_varlen(
 @pytest.mark.parametrize('H', test_h_list)
 @pytest.mark.parametrize('D', test_d_list)
 @pytest.mark.parametrize('gate_logit_normalizer', test_gate_list)
-@pytest.mark.parametrize('dtype', [torch.float])
+@pytest.mark.parametrize('dtype', [torch.float16])
 @pytest.mark.parametrize('scale', [1, 0.1])
 @pytest.mark.skipif(
     os.getenv('SKIP_TEST_CHUNK_VARLEN') == '0',
